requirements.txt
flask==2.0.1
pandas==1.3.3
mysql-connector-python==8.0.26